refactor: refactored transaction components
