+++ packages/dev-frontend/src/components/Tooltip.tsx
@@ -1,64 +1,42 @@
-import React, { useState, useRef } from "react";
-import { VirtualElement } from "@popperjs/core";
-import { usePopper } from "react-popper";
-import { Card } from "theme-ui";
+import Tippy from "@tippyjs/react";
+import type { TippyProps } from "@tippyjs/react";
+import React from "react";
+import { Box, Card, Link } from "theme-ui";
+import { Icon } from "./Icon";
 
-export type Hoverable = {
-  onMouseOver: () => void;
-  onMouseOut: () => void;
-  ref: (instance: Element | VirtualElement | null) => void;
+export type TooltipProps = Pick<TippyProps, "placement"> & {
+  message: React.ReactNode;
+  link?: string;
+  children?: React.ReactNode;
 };
 
-export type TooltipProps<C> = {
-  children: C;
-  message: string;
-  placement?: "top" | "bottom" | "left" | "right";
-};
+export type LearnMoreLinkProps = Pick<TooltipProps, "link">;
 
-export function Tooltip<C extends React.ReactElement<Hoverable>>({
-  children,
-  message,
-  placement = "top"
-}: TooltipProps<C>) {
-  const event = useRef<"over" | "out">();
-  const [show, setShow] = useState(false);
-  const [referenceElement, setReferenceElement] = useState<Element | VirtualElement | null>();
-  const [popperElement, setPopperElement] = useState<HTMLElement | null>();
-
-  const { styles, attributes } = usePopper(referenceElement, popperElement, { placement });
-
+export const LearnMoreLink: React.FC<LearnMoreLinkProps> = ({ link }) => {
   return (
-    <>
-      {React.cloneElement(React.Children.only<C>(children), {
-        // Use a debounced onMouseOver/onMouseOut instead of onMouseEnter/onMouseLeave to
-        // work around https://github.com/facebook/react/issues/10109
-
-        onMouseOver: () => {
-          event.current = "over";
-
-          if (!show) {
-            setShow(true);
-          }
-        },
-
-        onMouseOut: () => {
-          event.current = "out";
-
-          setTimeout(() => {
-            if (event.current === "out") {
-              setShow(false);
-            }
-          }, 0);
-        },
-
-        ref: setReferenceElement
-      })}
-
-      {show && (
-        <Card variant="tooltip" ref={setPopperElement} style={styles.popper} {...attributes.popper}>
-          {message}
-        </Card>
-      )}
-    </>
+    <Link href={link} target="_blank">
+      Learn more <Icon size="xs" name="external-link-alt" />
+    </Link>
   );
-}
+};
+
+export const Tooltip = ({ children, message, placement = "top", link }: TooltipProps): JSX.Element => {
+  return (
+    <Tippy
+      interactive={true}
+      placement={placement}
+      content={
+        <Card variant="tooltip">
+          {message}
+          {link && (
+            <Box mt={1}>
+              <LearnMoreLink link={link} />
+            </Box>
+          )}
+        </Card>
+      }
+    >
+      <span>{children}</span>
+    </Tippy>
+  );
+};
