+++ packages/dev-frontend/src/components/Transaction.tsx
@@ -1,10 +1,8 @@
 import React, { useState, useContext, useEffect, useCallback } from "react";
-import { Flex, Text, Box } from "theme-ui";
 import { Provider, TransactionResponse, TransactionReceipt } from "@ethersproject/abstract-provider";
 import { hexDataSlice, hexDataLength } from "@ethersproject/bytes";
 import { defaultAbiCoder } from "@ethersproject/abi";
 
-import { buildStyles, CircularProgressbarWithChildren } from "react-circular-progressbar";
 import "react-circular-progressbar/dist/styles.css";
 
 import { EthersTransactionOverrides, EthersTransactionCancelledError } from "@liquity/lib-ethers";
@@ -12,33 +10,11 @@ import { SentLiquityTransaction as SentThresholdTransaction, LiquityReceipt as T
 
 import { useThreshold } from "../hooks/ThresholdContext";
 
-import { Icon } from "./Icon";
-import { Tooltip, TooltipProps, Hoverable } from "./Tooltip";
+import { Tooltip } from "./Tooltip";
+import type { TooltipProps } from "./Tooltip";
 import { checkTransactionCollateral } from "../utils/checkTransactionCollateral";
 
-const strokeWidth = 10;
-
-const circularProgressbarStyle = {
-  strokeLinecap: "butt",
-  pathColor: "white",
-  trailColor: "rgba(255, 255, 255, 0.33)"
-};
-
-const slowProgress = {
-  strokeWidth,
-  styles: buildStyles({
-    ...circularProgressbarStyle,
-    pathTransitionDuration: 30
-  })
-};
-
-const fastProgress = {
-  strokeWidth,
-  styles: buildStyles({
-    ...circularProgressbarStyle,
-    pathTransitionDuration: 0.75
-  })
-};
+import { TransactionStatus } from "./TransactionStatus";
 
 type TransactionIdle = {
   type: "idle";
@@ -90,7 +66,7 @@ type TransactionConfirmedOneShot = {
   collateral?: string;
 };
 
-type TransactionState =
+export type TransactionState =
   | TransactionIdle
   | TransactionFailed
   | TransactionWaitingForApproval
@@ -157,11 +133,11 @@ export type TransactionFunction = (
 type TransactionProps<C> = {
   id: string;
   tooltip?: string;
-  tooltipPlacement?: TooltipProps<C>["placement"];
+  tooltipPlacement?: TooltipProps["placement"];
   showFailure?: "asTooltip" | "asChildText";
   requires?: readonly (readonly [boolean, string])[];
   send: TransactionFunction;
-  version: string;
+    version: string;
   collateral: string;
   children: C;
 };
@@ -190,17 +166,7 @@ export const useTransactionFunction = (
     } catch (error) {
       if (hasMessage(error) && error.message.includes("User denied transaction signature")) {
         setTransactionState({ type: "cancelled", id, version, collateral, });
-      }
-      else if (hasMessage(error) && error.message.includes("nothing to liquidate")) {
-        setTransactionState({
-          type: "failed",
-          id,
-          error: new Error("There are no vaults under-collateralized to liquidate"),
-          version,
-          collateral,
-        });
-      }
-      else {
+      } else {
         console.error(error);
 
         setTransactionState({
@@ -217,7 +183,7 @@ export const useTransactionFunction = (
   return [sendTransaction, transactionState];
 };
 
-export function Transaction<C extends React.ReactElement<ButtonlikeProps & Hoverable>>({
+export function Transaction<C extends React.ReactElement<ButtonlikeProps>>({
   id,
   tooltip,
   tooltipPlacement,
@@ -235,11 +201,11 @@ export function Transaction<C extends React.ReactElement<ButtonlikeProps & Hover
     .filter(([requirement]) => !requirement)
     .map(([, reason]) => reason);
 
-  const isCollateralChecked = checkTransactionCollateral(
-    transactionState,
-    version,
-    collateral
-  );
+    const isCollateralChecked = checkTransactionCollateral(
+      transactionState,
+      version,
+      collateral
+    );
 
   if (
     isCollateralChecked &&
@@ -294,53 +260,9 @@ const tryToGetRevertReason = async (provider: Provider, tx: TransactionReceipt) 
   }
 };
 
-const Donut = React.memo(
-  CircularProgressbarWithChildren,
-  ({ value: prev }, { value: next }) => prev === next
-);
-
-type TransactionProgressDonutProps = {
-  state: TransactionState["type"];
-};
-
-const TransactionProgressDonut = ({ state }: TransactionProgressDonutProps): JSX.Element => {
-  const [isMounted, setIsMounted] = useState<boolean>(true);
-  const [value, setValue] = useState(0);
-  const maxValue = 1;
-
-  useEffect(() => {
-    if (!isMounted) {
-      return
-    }
-
-    if (state === "confirmed") {
-      setTimeout(() => setValue(maxValue), 40);
-    } else {
-      setTimeout(() => setValue(maxValue * 0.67), 20);
-    }
-
-    return () => { 
-      setIsMounted(false);
-    };
-  }, [state, isMounted]);
-
-  return state === "confirmed" ? (
-    <Donut {...{ value, maxValue, ...fastProgress }}>
-      <Icon name="check" color="white" size="lg" />
-    </Donut>
-  ) : state === "failed" || state === "cancelled" ? (
-    <Donut value={0} {...{ maxValue, ...fastProgress }}>
-      <Icon name="times" color="white" size="lg" />
-    </Donut>
-  ) : (
-    <Donut {...{ value, maxValue, ...slowProgress }}>
-      <Icon name="cog" color="white" size="lg" spin />
-    </Donut>
-  );
-};
-
 export const TransactionMonitor = (): JSX.Element => {
   const [isMounted, setIsMounted] = useState<boolean>(true);
+
   const { provider } = useThreshold();
   const [transactionState, setTransactionState] = useTransactionState();
 
@@ -437,7 +359,7 @@ export const TransactionMonitor = (): JSX.Element => {
 
   useEffect(() => {
     if (!isMounted) return;
-    
+
     if (transactionState.type === "confirmedOneShot" && id) {
       // hack: the txn confirmed state lasts 5 seconds which blocks other states, review with Dani
       setTransactionState({ type: "confirmed", id });
@@ -467,39 +389,9 @@ export const TransactionMonitor = (): JSX.Element => {
   }
 
   return (
-    <Flex
-      sx={{
-        alignItems: "center",
-        bg:
-          transactionState.type === "confirmed"
-            ? "success"
-            : transactionState.type === "cancelled"
-            ? "warning"
-            : transactionState.type === "failed"
-            ? "danger"
-            : "primary",
-        p: 3,
-        pl: 4,
-        left: 0,
-        position: "fixed",
-        width: "100vw",
-        bottom: 0,
-        overflow: "hidden"
-      }}
-    >
-      <Box sx={{ mr: 3, width: "40px", height: "40px" }}>
-        <TransactionProgressDonut state={transactionState.type} />
-      </Box>
-
-      <Text sx={{ fontSize: 3, color: "white" }}>
-        {transactionState.type === "waitingForConfirmation"
-          ? "Waiting for confirmation"
-          : transactionState.type === "cancelled"
-          ? "Cancelled"
-          : transactionState.type === "failed"
-          ? transactionState.error.message
-          : "Confirmed"}
-      </Text>
-    </Flex>
+    <TransactionStatus
+      state={transactionState.type}
+      message={transactionState.type === "failed" ? transactionState.error.message : undefined}
+    />
   );
 };
